+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## (Unreleased)
 
+* Changed: Reduced the scope of the Anonymizer#anonymize method (introduced in 0.9.5) from files, folders & objects to only DICOM objects.
+* Added: Anonymizer#anonymize_path method for anonymizing files or directories. This method is able to anonymize thousands of files without any memory issues.
 * Fixed: An element of VR 'UN' with length -1 is treated as a sequence.
 * Fixed: Properly handle private sequence with implicit data inside and explicit file.
 * Fixed: Set license type according to spdx.org.
 * Fixed: Calculation of field length in cases where negotiation of Transfer Syntax fails.
 * Update: The DICOM Dictionary where updated to the current version (2016).
+
 
 # 0.9.6